+++ simple-stack-flow-masterdetail-fragments/src/main/java/com/example/stackmasterdetailfrag/MainActivity.java
@@ -29,12 +29,8 @@ import com.example.stackmasterdetailfrag.util.FragmentManagerService;
 import com.example.stackmasterdetailfrag.util.MasterDetailBackstackDelegate;
 import com.zhuinden.simplestack.BackstackDelegate;
 import com.zhuinden.simplestack.HistoryBuilder;
-import com.zhuinden.simplestack.ServiceFactory;
 import com.zhuinden.simplestack.StateChange;
 import com.zhuinden.simplestack.StateChanger;
-
-import java.util.List;
-import java.util.Map;
 
 import static android.view.MenuItem.SHOW_AS_ACTION_ALWAYS;